Добавил валидацию по запросу, cохранение и подстановка значений из последнего запроса,и создал класс

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Cache\Events\CacheEvent;
 use Illuminate\Http\Request;
@@ -30,11 +31,15 @@ class CategoryController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+       if ($request->has('image')){
+           $path = $request->file('image')->store('categories');
+
+           $params['image'] = $path;
+       }
         Category::create($params);
         return redirect()->route('categories.index');
     }
@@ -59,14 +64,18 @@ class CategoryController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
-        Category::create($params);
-        $category->update($params);
+        unset($params['image']);
+        if ($request->has('image')){
+            Storage::delete($category->image);
+            $path = $request->file('image')->store('categories');
+            $params['image'] = $path;
+            Category::create($params);
+            $category->update($params);
+        }
+        return redirect()->route('categories.index');
     }
 
     /**
